// src/app/pages/model/administrator/system/Currency.ts
export interface Currency {
    id?: number | undefined;
    currencyCode?: string;
    currencyChar?: string;
    currencyNumber?: string;
    currencyName?: string;
    localCurrencyName?: string;
    currencySymbol?: string;
    decimalDigits?: boolean;
    roundingDigits?: boolean;
    displayOrder?: boolean;
    currencyStatus?: string;
    description: string;
}